bluetooth-audio: Python 3 support / use separate dbus process
