+++ packages/addons/service/bluetooth-audio/changelog.txt
@@ -1,3 +1,7 @@
+102
+- Python 3 support
+- Using separate dbus process
+
 101
 - Fix log errors
